+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSRuinsConfig.java
@@ -23,13 +23,13 @@ public class RSRuinsConfig
 					.comment("How rare are Warm Land Ruins in Plains, Forests, Swamps, and non-snowy Taiga biomes.",
 							"\n1 for spawning in most chunks and 1001 for none.")
 					.translation("repurposedstructures.config.ruins.ruinslandwarmmaxchunkdistance")
-					.defineInRange("ruinsLandWarmMaxChunkDistance", 36, 1, 1001));
+					.defineInRange("ruinsLandWarmMaxChunkDistance", 42, 1, 1001));
 
 			ruinsLandHotMaxChunkDistance = subscriber.subscribe(builder
 					.comment("How rare are Hot Land Ruins in Desert biomes.",
 							"\n1 for spawning in most chunks and 1001 for none.")
 					.translation("repurposedstructures.config.ruins.ruinslandhotmaxchunkdistance")
-					.defineInRange("ruinsLandHotMaxChunkDistance", 39, 1, 1001));
+					.defineInRange("ruinsLandHotMaxChunkDistance", 45, 1, 1001));
 		}
 	}
 }