+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/features/StructureEndRodChains.java
@@ -4,18 +4,20 @@ import com.mojang.serialization.Codec;
 import com.telepathicgrunt.repurposedstructures.world.features.configs.StructureTargetConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.server.level.WorldGenRegion;
 import net.minecraft.world.level.WorldGenLevel;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.EndRodBlock;
 import net.minecraft.world.level.block.LanternBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 
 
-public class StructureChains extends Feature<StructureTargetConfig> {
+public class StructureEndRodChains extends Feature<StructureTargetConfig> {
 
-    public StructureChains(Codec<StructureTargetConfig> config) {
+    public StructureEndRodChains(Codec<StructureTargetConfig> config) {
         super(config);
     }
 
@@ -28,9 +30,9 @@ public class StructureChains extends Feature<StructureTargetConfig> {
 
         for(int i = 0; i < context.config().attempts; i++){
             mutable.set(context.origin()).move(
-                    context.random().nextInt(11) - 5,
-                    context.random().nextInt(3) - 1,
-                    context.random().nextInt(11) - 5
+                    context.random().nextInt(7) - 3,
+                    -1,
+                    context.random().nextInt(7) - 3
             );
 
             if(!world.getBlockState(mutable).isAir()){
@@ -39,14 +41,14 @@ public class StructureChains extends Feature<StructureTargetConfig> {
 
             // generates chains from given position down 1-8 blocks if path is clear and the given position is valid
             int length = 0;
-            BlockState aboveBlockstate;
+            BlockState belowBlockstate;
             boolean exitEarly = false;
 
-            for (; mutable.getY() > world.getMinBuildHeight() + 3 && length < context.random().nextInt(context.random().nextInt(context.random().nextInt(8) + 1) + 1) + 1; mutable.move(Direction.DOWN)) {
+            for (; mutable.getY() < world.getMaxBuildHeight() - 3 && length < context.random().nextInt(context.random().nextInt(context.random().nextInt(8) + 1) + 1) + 1; mutable.move(Direction.UP)) {
                 if (world.isEmptyBlock(mutable)) {
-                    aboveBlockstate = world.getBlockState(mutable.above());
+                    belowBlockstate = world.getBlockState(mutable.below());
 
-                    if (aboveBlockstate.is(Blocks.CHAIN) || aboveBlockstate.isFaceSturdy(world, mutable.above(), Direction.DOWN)) {
+                    if (belowBlockstate.is(Blocks.CHAIN) || belowBlockstate.isFaceSturdy(world, mutable.below(), Direction.DOWN)) {
                         world.setBlock(mutable, Blocks.CHAIN.defaultBlockState(), 2);
                         length++;
                     }
@@ -58,14 +60,13 @@ public class StructureChains extends Feature<StructureTargetConfig> {
 
             if(exitEarly) continue;
 
-            //attaches lantern at end at a rare chance
-            if(mutable.getY() != world.getMinBuildHeight() + 3 && context.random().nextFloat() < 0.075f && world.isEmptyBlock(mutable)){
-                if(world.getBiome(mutable).getBiomeCategory() == Biome.BiomeCategory.NETHER){
-                    world.setBlock(mutable, Blocks.SOUL_LANTERN.defaultBlockState().setValue(LanternBlock.HANGING, true), 2);
-                }
-                else{
-                    world.setBlock(mutable, Blocks.LANTERN.defaultBlockState().setValue(LanternBlock.HANGING, true), 2);
-                }
+            //attaches end rod at end at a decent chance
+            if(mutable.getY() != world.getMaxBuildHeight() - 3 &&
+                context.random().nextFloat() < 0.15f &&
+                world.isEmptyBlock(mutable.above()) &&
+                world.isEmptyBlock(mutable))
+            {
+                world.setBlock(mutable, Blocks.END_ROD.defaultBlockState().setValue(EndRodBlock.FACING, Direction.UP), 2);
             }
         }